pyproject.toml
[project]
name = "temporary-mcp-servers"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "aiohttp~=3.13.1",
    "dotenv-pth~=1.0",
    "fake-useragent~=2.2.0",
    "ipython~=9.6.0",
    "logfire[aiohttp-client]~=4.14.1",
    "mcp-hmr~=0.0.2.3",
    "objprint~=0.3.0",
    "python-readability~=1.0.0rc0",
    "pywebview~=6.1",
    "saneyaml~=0.6.1",
    "stamina~=25.1.0",
]

[tool.pyright]
typeCheckingMode = "standard"
reportUnnecessaryTypeIgnoreComment = "error"

[tool.ruff]
line-length = 200

[tool.ruff.lint]
extend-select = [
    "FURB",  # refurb
    "RUF",   # ruff
    "UP",    # pyupgrade
    "N",     # pep8-naming
    "W",     # pycodestyle
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "FA",    # flake8-future-annotations
    "TC",    # flake8-type-checking
    "FBT",   # flake8-boolean-trap
    "ISC",   # flake8-implicit-str-concat
    "ARG",   # flake8-unused-arguments
    "PIE",   # flake8-pie
    "PTH",   # flake8-use-pathlib
    "RSE",   # flake8-raise
    "SIM",   # flake8-simplify
    "SLF",   # flake8-self
    "ASYNC", # flake8-async
]
